:truck: move ExtensionPanelContent
