+++ app/components/Package/index.js
@@ -1,3 +1,2 @@
 /* @flow weak */
 export PackageControlView from './PackageControlView'
-export ExtensionPanelContent from './ExtensionPanelContent'
