+++ app/components/Panel/ExtensionPanelContent.js
